Refactored the links component to match the implementation of board authentication. Added info message to communicate how passwords work with the boards in the new auth system.
